requirements-fixed.txt
dbt-core==1.6.1
dbt-duckdb==1.6.0
snowplow-tracker==0.8.0
markupsafe==2.0.1
click==8.0.4
